fix: pids are 64 bit

+++ src/ipc.rs
@@ -4,7 +4,7 @@ pub struct Handle(pub u64);
 #[repr(C)]
 pub struct IpcMessageHeader {
     pub length: u32,   
-    pub sender_pid: u32,
+    pub sender_pid: u64,
     pub reply_handle: u64,
 }
